[GH-1390] Remove redundant calls to Fog.credentials
If you use the supported means to create services such as:

    Fog::Service.new(:provider => "Example")

It already loads the contents of `~/.fog` and filters the values based
on if they are recognised or required.

That means in a number of services that are initialising based on
`options` or a value from `Fog.credentials` both are the same.

The `options` should have been loaded and merged from credentials in the
service "builder".